+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Unreleased
 - Added `**kwargs` to `AzureBlobFileSystem.exists()`
 - Populate `AzureBlobFile.version_id` on write when `version_aware` is enabled.
 - Fixed issue where unawaitable Credential types were incorrectly awaited (#431)
-- Added `AzureBlobFileSystem.mv()` to address issue of files not being deleted when they are moved.
+- Fixed a bug where moving files does not delete them from the original location
 
 2026.2.0
 --------